(#232) GitHub split and __init__
